작업 PR

+++ booking/src/main/java/rice_monkey/booking/feign/listing/dto/ListingDto.java
@@ -3,7 +3,7 @@ package rice_monkey.booking.feign.listing.dto;
 public record ListingDto(
         Long id,
         String name,
-        int price,
+        long price,
         String status
 ) {
 }
